Web/Notepad--/Challenge/Private/src/requirements.txt
Flask
Flask-Login
SQLAlchemy
gunicorn
pytest-playwright
